+++ lib/src/actors/libspiffy_actor_system.dart
@@ -623,6 +623,8 @@ class LibSpiffyActorSystem {
         aliases: const ['PaymentAcknowledgedEvent']);
     EventRegistry.register<PaymentCountersignedEvent>(PaymentCountersignedEvent.stableTypeName, PaymentCountersignedEvent.fromMap,
         aliases: const ['PaymentCountersignedEvent']);
+    EventRegistry.register<ReturnLegRecordedInWalletEvent>(ReturnLegRecordedInWalletEvent.stableTypeName, ReturnLegRecordedInWalletEvent.fromMap,
+        aliases: const ['ReturnLegRecordedInWalletEvent']);
     EventRegistry.register<ChannelClosingEvent>(ChannelClosingEvent.stableTypeName, ChannelClosingEvent.fromMap,
         aliases: const ['ChannelClosingEvent']);
     EventRegistry.register<ChannelClosedEvent>(ChannelClosedEvent.stableTypeName, ChannelClosedEvent.fromMap,