+++ test/core/event_type_names_test.dart
@@ -70,6 +70,7 @@ const goldenTypeNames = <String, String>{
   'PaymentRecordedEvent': 'channel.payment.recorded',
   'PaymentAcknowledgedEvent': 'channel.payment.acknowledged',
   'PaymentCountersignedEvent': 'channel.payment.countersigned',
+  'ReturnLegRecordedInWalletEvent': 'channel.return_leg.wallet_recorded',
   'ChannelClosingEvent': 'channel.closing',
   'ChannelClosedEvent': 'channel.closed',
   'RefundClaimedEvent': 'channel.refund.claimed',
@@ -274,6 +275,8 @@ Map<String, Event> sampleEvents() => <String, Event>{
           channelId: _c, sequenceNumber: 1, serverSignatureHex: '30',
           fullySignedPaymentTxHex: '01', fullySignedPaymentTxId: 'ab' * 32,
           timestamp: _t, version: 9),
+      'ReturnLegRecordedInWalletEvent': ReturnLegRecordedInWalletEvent(
+          channelId: _c, txId: 'cd' * 32, timestamp: _t, version: 9),
       'ChannelClosingEvent': ChannelClosingEvent(
           channelId: _c, initiator: 'client',
           clientBalanceSats: BigInt.from(890),